fix: unify npm and plugin versions into one number
The plugin family (.claude-plugin/plugin.json and friends) and the npm
package used two independent version numbers bumped by separate scripts
(release vs release:plugin). Nothing enforced running both together, so
the plugin manifest drifted to 5.2.1 while npm moved on to 6.2.10 —
claude plugin update then reported "already at the latest version" even
though the repo had shipped since (#393).

Collapses to a single version carried by package.json and five other
files (src/hosts/gemini/gemini-extension.json, the four plugin-family
manifests), all bumped together by one npm run release. This removes the
possibility of the two drifting apart rather than just re-syncing them
once.
