+++ .cursor-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "name": "hedgehog",
   "description": "Offers to set up the Hedgehog build discipline in any project you open",
-  "version": "5.2.1",
+  "version": "6.2.10",
   "author": {
     "name": "skyf0xx"
   },
